refactor: change name of file to just sokoban
